+++ packages/orq-rc/RELEASES.md
@@ -4199,3 +4199,13 @@ Based on:
 - [typescript v3.14.0-rc.20] packages/orq-rc
 ### Releases
 - [NPM v3.14.0-rc.20] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.20 - packages/orq-rc
+
+## 2025-10-12 11:49:25
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.636.3 (2.723.11) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v3.14.0-rc.21] packages/orq-rc
+### Releases
+- [NPM v3.14.0-rc.21] https://www.npmjs.com/package/@orq-ai/node/v/3.14.0-rc.21 - packages/orq-rc